handle case where fibonacci values are greater than size of integer

+++ src/main/java/org/example/Factorial.java
@@ -1,7 +1,7 @@
 package org.example;
 
 public class Factorial {
-    static int factorial(int n) {
+    static long factorial(int n) {
         if (n == 0) {
             return 1;
         }
